Allowed storing non-integer score target multipliers between -1 and 1

+++ app/controllers/checklists_controller.rb
@@ -120,7 +120,7 @@ class ChecklistsController < ApplicationController
     newTopics.each do |idkey, values|
       # First we fix the scale rational values
       target = BigDecimal(values["score_target"])
-      unless values["scale_denominator"].to_i == 0 or values["scale_numerator"].to_i == 0 
+      unless values["scale_denominator"].to_f == 0 or values["scale_numerator"].to_f == 0 
         scale = Rational(values["scale_numerator"], values["scale_denominator"]).round(5)
       else
         scale = Rational(1)
